Move the ssg connections from utils to cac transformer
