+++ trestlebot/cli/utils.py
@@ -1,17 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright (c) 2024 Red Hat, Inc.
 
-import logging
 from typing import Any, Dict, List
-
-from ssg.products import load_product_yaml, product_yaml_path
 
 from trestlebot.bot import TrestleBot
 from trestlebot.reporter import BotResults
 from trestlebot.tasks.base_task import TaskBase
-
-
-logger = logging.getLogger(__name__)
 
 
 def comma_sep_to_list(string: str) -> List[str]:
@@ -41,16 +35,3 @@ def run_bot(pre_tasks: List[TaskBase], kwargs: Dict[Any, Any]) -> BotResults:
         ),
         dry_run=kwargs.get("dry_run", False),
     )
-
-
-def get_component_title(product_name: str, cac_path: str) -> str:
-    """Get the product name from product yml file via the SSG library."""
-    if product_name and cac_path:
-        # Get the product yaml file path
-        product_yml_path = product_yaml_path(cac_path, product_name)
-        # Load the product data
-        product = load_product_yaml(product_yml_path)
-        # Return product name from product yml file
-        return product._primary_data.get("product")
-    else:
-        raise ValueError("component_title is empty or None")
